refactor(player): change the rect morph animation to play sequentially rather than parallel to the element fade-in and fade-out animations
When transitioning slides, first the fade-out animation is played, then the morph animation, and finally the fade-in animation.

This change was made because the morph animations change the order of the elements, and with it playing alongside the fade-in/fade-out animations, some elements would get hidden during the slide transition and their fade-out animations would not be visible.