improvement(generation): when separating a paragraphs rect, balance the resulting paragraph rects to have almost the same area of empty space
This change aims to prevent a big paragraph from getting squished inside its container when there's a small paragraph with a lot of empty area. This also removes the point of randomizing the size of each paragraph rect, so maybe this balancing should only happen in the situation just described.